refactor: packages stream layout
