+++ video-sdk/src/main/java/com/kaleyra/video_sdk/call/stream/layoutsystem/itemsprovider/MosaicStreamItemsProvider.kt
@@ -1,7 +1,7 @@
 package com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider
 
-import com.kaleyra.video_sdk.call.stream.model.StreamItem
-import com.kaleyra.video_sdk.call.stream.model.MoreStreamsUserPreview
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.StreamItem
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.MoreStreamsUserPreview
 import com.kaleyra.video_sdk.call.stream.model.core.StreamUi
 
 /**
