+++ video-sdk/src/main/java/com/kaleyra/video_sdk/call/stream/layoutsystem/layout/ManualLayout.kt
@@ -1,8 +1,8 @@
 package com.kaleyra.video_sdk.call.stream.layoutsystem.layout
 
 import com.kaleyra.video_sdk.call.stream.layoutsystem.config.StreamLayoutConstraints
-import com.kaleyra.video_sdk.call.stream.model.StreamItem
-import com.kaleyra.video_sdk.call.stream.model.StreamItemState
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.StreamItem
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.StreamItemState
 import com.kaleyra.video_sdk.call.stream.model.core.StreamUi
 import com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider.FeaturedStreamItemsProvider
 import com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider.FeaturedStreamItemsProviderImpl
