+++ video-sdk/src/test/java/com/kaleyra/video_sdk/layoutsystem/itemsprovider/FeaturedStreamItemsProviderTest.kt
@@ -1,10 +1,10 @@
-package com.kaleyra.video_sdk.layout
+package com.kaleyra.video_sdk.layoutsystem.itemsprovider
 
-import com.kaleyra.video_sdk.call.stream.model.StreamItem
-import com.kaleyra.video_sdk.call.stream.model.MoreStreamsUserPreview
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.StreamItem
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.MoreStreamsUserPreview
 import com.kaleyra.video_sdk.call.stream.model.core.StreamUi
 import com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider.FeaturedStreamItemsProviderImpl
-import com.kaleyra.video_sdk.call.stream.model.StreamItemState
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.StreamItemState
 import com.kaleyra.video_sdk.common.avatar.model.ImmutableUri
 import io.mockk.mockk
 import junit.framework.TestCase.assertEquals
