+++ video-sdk/src/test/java/com/kaleyra/video_sdk/layoutsystem/model/StreamItemTest.kt
@@ -1,9 +1,9 @@
-package com.kaleyra.video_sdk.layout
+package com.kaleyra.video_sdk.layoutsystem.model
 
-import com.kaleyra.video_sdk.call.stream.model.StreamItem
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.StreamItem
 import com.kaleyra.video_sdk.call.stream.model.core.StreamUi
 import com.kaleyra.video_sdk.call.stream.model.core.VideoUi
-import com.kaleyra.video_sdk.call.stream.model.StreamItemState
+import com.kaleyra.video_sdk.call.stream.layoutsystem.model.StreamItemState
 import org.junit.Assert
 import org.junit.Test
